New set operation implementation started everything compiles
There is a new set operation trait for graph_obj.
That unifies the name of the operations.
The heavy lifting is done under the ops/graph_obj/setops.rs
A reference based approach of the function under there could help.
Or some other thing that is more sensitive to ownership.