Path and Tree implementations started
Path and Tree traits are written. There are some methods that need
to be translated to support various algorithms from python code.
But it shouldn't prove too difficult. One may need to handle
graph searcher first though due to need for topological sort.
Something like bfs or dfs handler might be necessary. There is a
specific module for this in Python code as well.